backend/requirements.txt
fastapi==0.109.*
uvicorn[standard]
openai
python-dotenv
langdetect>=1.0
playwright
beautifulsoup4
lxml
requests
tiktoken
chromadb
sseclient-py
scikit-learn
numpy          # pulled by scikit-learn, but pin explicitly for reproducibility
scipy          # same as above
gradio==3.50.*
pydantic==1.10.15
sendgrid==6.11.0
google-auth
google-auth-oauthlib
google-api-python-client
flask>=3.0       
twilio>=9.0
langdetect>=1.0
